general-selectors are TS
